+++ packages/core/index.js
@@ -1,5 +1,0 @@
-import { execa } from 'execa'
-
-const result = await execa('pnpm', ['root', '-g'])
-
-console.log(result)