+++ packages/core/src/modules/util.ts
@@ -2,6 +2,8 @@
  * 常用工具模块
  * @author: sunkeysun
  */
+import { filesize as fz } from 'filesize'
+
 export function capitalize(str: string) {
   return str.replace(
     /^(\w)(\w*)/,
@@ -15,4 +17,8 @@ export function getPluginName(packageName: string) {
     .replace('@inventorjs/plugin-', '')
     .replace(/^(@[\w-_]+\/)?inventor-plugin-/g, '')
   return pluginName
+}
+
+export function humanSize(bytes: number, options = { base: 2, standard: 'jedec' }) {
+  return fz(bytes, options)
 }