+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @inventorjs/cli
 
+## 3.0.2
+
+### Patch Changes
+
+- a537206: change loading task
+- Updated dependencies [a537206]
+  - @inventorjs/core@3.11.2
+  - @inventorjs/plugin-app@3.0.2
+  - @inventorjs/plugin-plugin@3.0.2
+
 ## 3.0.1
 
 ### Patch Changes